requirements.txt
ccxt
pandas
numpy
python-telegram-bot
